+++ src/tiles/grid/item-tile.ts
@@ -10,7 +10,7 @@ import { reviewsIcon } from './icons/reviews';
 import viewsIcon from './icons/views';
 
 import '../mediatype-icon';
-import '../item-image';
+import '../item-tile-image';
 
 @customElement('item-tile')
 export class ItemTile extends LitElement {
@@ -27,8 +27,11 @@ export class ItemTile extends LitElement {
         <div id="title-image-container">
           <h1 id="item-title" title=${itemTitle}>${this.model?.title}</h1>
           <div id="item-image-container">
-            <item-image .model=${this.model} .baseImageUrl=${this.baseImageUrl}>
-            </item-image>
+            <item-tile-image
+              .model=${this.model}
+              .baseImageUrl=${this.baseImageUrl}
+            >
+            </item-tile-image>
           </div>
           <div class="item-creator">
             <div class="truncated">
